Rename src/util.ts to src/utils.ts
