+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './interfaces';
 export * from './Managers';
 export * from './Measures';
-export * from './util';
+export * from './utils';
 export * from './measure_visualization/ParallelSets';
 export * from './measure_visualization/BoxPlot';
 export * from './measure_visualization/ScatterPlot';
